refactor: group image files
